refactor: changed Fold to provide receiver
Choosing to prefer the receiver due to the added benefit that it "unwraps" its internal property as a named variable. It also makes most lambda references more descriptive (albeit a bit harder to write), although some values that were previously references had to become lambdas because there was no reference matching the new signature that could utilise the internal value.